requirements.txt
discord.py==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
PyMuPDF==1.23.19
